Add a string literal that is passed verbatim to CSS
Sometimes CleverCSS is a bit too clever for its own good and you just
want to pass something directly to CSS. For instance, functions that
aren't rgb() or url() need to be escaped. Added is a simple new
backtick-surrounded string format that will be passed verbatim without
further processing. Example::

  .gradient:
    background: `-moz-linear-gradient(...)`